feat(signup events): add Subscriber class to hold the signup logic and events
Added the Subscriber class to hold and handle the signup logic. Added the event hookd for before
signup and after signup, Renamed SendEvent to MailEvent be be consistent with the Controller naming
