+++ composer.json
@@ -2,7 +2,7 @@
     "name": "wc4bcraftsibcontactform/craft-sib-contact-form",
     "description": "A contact form integration for the Send in Blue API",
     "type": "craft-plugin",
-    "version": "1.2.0",
+    "version": "1.3.0",
     "keywords": [
         "craft",
         "cms",
@@ -40,7 +40,8 @@
         "changelogUrl": "https://raw.githubusercontent.com/craft-sib-contact-form/craft-sib-contact-form/master/CHANGELOG.md",
         "class": "wc4bcraftsibcontactform\\craftsibcontactform\\Craftsibcontactform",
         "components": {
-            "mailer": "wc4bcraftsibcontactform\\craftsibcontactform\\Mailer"
+            "mailer": "wc4bcraftsibcontactform\\craftsibcontactform\\Mailer",
+            "subscriber": "wc4bcraftsibcontactform\\craftsibcontactform\\Subscriber"
         }
     }
 }
